requirements.txt
hydra-core
hydra-colorlog
torch
torchvision
torchmetrics
lightning
transformers
numpy
matplotlib
opencv-python
tqdm
wandb
timm==0.9.10
mmh3

